requirements.txt
discord.py==2.3.1
langchain==0.0.217
matplotlib==3.7.1
openai==0.27.8
pandas==2.0.0
